refactor to different classes
