+++ src/frontend/src/Pages/HomePage.js
@@ -2,19 +2,13 @@ import React, { useEffect, useState } from 'react';
 import { useHistory, withRouter } from 'react-router-dom';
 import axios from 'axios';
 
-import { Typography, Row, Col, Divider, Button } from "antd";
-import { PlusOutlined } from "@ant-design/icons";
-import ReactPlayer from 'react-player'
-import SlideCard from '../Components/SlideCard/SliceCard';
-import ThreeDotsMenu from '../Components/HomePage/ThreeDotMenu';
+import IntroView from '../Components/HomePage/IntroView';
+import StoriesView from '../Components/HomePage/StoriesView';
+import NewStoryButton from '../Components/HomePage/CreateNewStory';
 import "antd/dist/antd.css";
 
-const { Title, Text } = Typography;
-
-const style = { padding: "0px" };
 
 const homepage_style = { "padding": "20px", direction: "rtl" };
-const homepage_card_style = { "margin": "0px", "width": "100%" };
 
 const HomePage = withRouter((props) => {
     const [stories, setStories] = useState([]);
@@ -29,37 +23,11 @@ const HomePage = withRouter((props) => {
     }, []);
     console.log(stories);
     return (
-        <div offset={4} style={homepage_style}>
-            <Divider orientation="right"><Title level={2}>הסיפורים שלי</Title></Divider>
-            <Row justify="end">
-                <a href="#" style={{ padding: "8px" }}>כל הסיפורים</a>
-            </Row>
-            <Row gutter={16}>
-                {stories.map(story => (
-                    <Col className="gutter-row" span={6}>
-                        <SlideCard cardProps={{ onClick: () => history.push({ pathname: `/story/${story.storyId}` }) }} cardStyle={homepage_card_style} imageUrl={story.thumbnail} />
-                        <ThreeDotsMenu></ThreeDotsMenu>
-                    </Col>
-                ))}
-            </Row>
-            <div style={{ "padding-top": "40px", "padding-bottom": "30px", direction: "rtl" }}>
-                <Button onClick={() => { history.push({ pathname: "/create" }) }} type="primary" shape="round" icon={<PlusOutlined />} size={"large"}>
-                    צור סיפור חדש
-            </Button>
-            </div>
-
+        <div offset={4} style={homepage_style}> 
+            <StoriesView stories={stories} history={history}></StoriesView>
+            <NewStoryButton history={history}></NewStoryButton>
             {/* The URL should be replaced with the tutorial video for the system  */}
-            <Divider orientation="right"><Title level={3}>איך זה עובד?</Title></Divider>
-            <Row gutter={16}>
-                <Col className="gutter-row" span={8}>
-                    <ReactPlayer
-                        url='https://www.youtube.com/watch?v=ysz5S6PUM-U'
-                        light={true}
-                        playing
-                        controls
-                    />
-                </Col>
-            </Row>
+            <IntroView url='https://www.youtube.com/watch?v=ysz5S6PUM-U'></IntroView>;
         </div>
     )
 });
